pyproject.toml
[project]
name = "codeactagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.5.5",
    "litellm>=1.70.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
]
